% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-core.R
\name{check_dockerfile}
\alias{check_dockerfile}
\title{Ensure an object is a dockerfile}
\usage{
check_dockerfile(dockerfile)
}
\arguments{
\item{dockerfile}{Object to check}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
Ensure an object is a dockerfile
}
