% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-core.R
\name{check_dockerignore}
\alias{check_dockerignore}
\title{Ensure an object is a dockerignore}
\usage{
check_dockerignore(dockerignore)
}
\arguments{
\item{dockerignore}{Object to check}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
Ensure an object is a dockerignore
}
