% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_linux_distribution}
\alias{determine_linux_distribution}
\title{Determine the Linux Distribution from a base image}
\usage{
determine_linux_distribution(base_image)
}
\arguments{
\item{base_image}{Base image name}
}
\value{
Character string of OS type (e.g., "ubuntu", "centos", "debian", "alpine")
}
\description{
Analyzes a Docker base image name to determine the underlying distribution
}
\details{
This function parses the base image name to extract the underlying
distribution. For Rocker Project images (which are based on Ubuntu/Debian),
it returns "ubuntu". For other distributions, it attempts to identify
common ones like Debian, CentOS, Fedora, Alpine, etc.
}
\examples{
determine_linux_distribution("rocker/r-ver:4.4.0")  # Returns "ubuntu"
determine_linux_distribution("alpine:3.16")         # Returns "alpine"
determine_linux_distribution("centos:7")            # Returns "centos"

}
\seealso{
\code{\link[=determine_package_manager]{determine_package_manager()}} for determining the package manager &
\code{\link[=map_to_sysreqs_platform]{map_to_sysreqs_platform()}} for mapping to sysreqs platform

Other utility functions: 
\code{\link{determine_package_manager}()},
\code{\link{map_to_sysreqs_platform}()}
}
\concept{utility functions}
