% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_os}
\alias{determine_os}
\title{Determine the operating system from a base image}
\usage{
determine_os(base_image)
}
\arguments{
\item{base_image}{Base image name}
}
\value{
Character string of OS type (e.g., "ubuntu", "centos", "debian")
}
\description{
Determine the operating system from a base image
}
