% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_package_manager}
\alias{determine_package_manager}
\title{Determine the package manager from a base image}
\usage{
determine_package_manager(base_image)
}
\arguments{
\item{base_image}{Base image name}
}
\value{
Character string of package manager type ("apt", "yum", "apk", etc.)
}
\description{
Determine the package manager from a base image
}
