% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_cmd}
\alias{dfi_cmd}
\title{Add a CMD instruction to a dockerfile}
\usage{
dfi_cmd(dockerfile, command)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{command}{Command to run}
}
\value{
Updated dockerfile object
}
\description{
Add a CMD instruction to a dockerfile
}
