% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_entrypoint}
\alias{dfi_entrypoint}
\title{Add an ENTRYPOINT instruction to a dockerfile}
\usage{
dfi_entrypoint(dockerfile, entrypoint)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{entrypoint}{Entrypoint command}
}
\value{
Updated dockerfile object
}
\description{
Add an ENTRYPOINT instruction to a dockerfile
}
