% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_expose}
\alias{dfi_expose}
\title{Add an EXPOSE instruction to a dockerfile}
\usage{
dfi_expose(dockerfile, ports)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{ports}{Ports to expose (numeric or character)}
}
\value{
Updated dockerfile object
}
\description{
Add an EXPOSE instruction to a dockerfile
}
