% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_from}
\alias{dfi_from}
\title{Add a FROM instruction to a dockerfile}
\usage{
dfi_from(dockerfile, image, as = NULL)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{image}{Base image name}

\item{as}{Name for this build stage (optional)}
}
\value{
Updated dockerfile object
}
\description{
Add a FROM instruction to a dockerfile
}
