% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_label}
\alias{dfi_label}
\title{Add a LABEL instruction to a dockerfile}
\usage{
dfi_label(dockerfile, ...)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{...}{Named labels}
}
\value{
Updated dockerfile object
}
\description{
Add a LABEL instruction to a dockerfile
}
