% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_maintainer}
\alias{dfi_maintainer}
\title{Add a MAINTAINER instruction to a dockerfile (deprecated)}
\usage{
dfi_maintainer(dockerfile, maintainer)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{maintainer}{Maintainer info}
}
\value{
Updated dockerfile object
}
\description{
Add a MAINTAINER instruction to a dockerfile (deprecated)
}
