% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_onbuild}
\alias{dfi_onbuild}
\title{Add an ONBUILD instruction to a dockerfile}
\usage{
dfi_onbuild(dockerfile, instruction)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{instruction}{Instruction to run on build}
}
\value{
Updated dockerfile object
}
\description{
Add an ONBUILD instruction to a dockerfile
}
