% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_shell}
\alias{dfi_shell}
\title{Add a SHELL instruction to a dockerfile}
\usage{
dfi_shell(dockerfile, shell)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{shell}{Shell command and parameters}
}
\value{
Updated dockerfile object
}
\description{
Add a SHELL instruction to a dockerfile
}
