% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_stopsignal}
\alias{dfi_stopsignal}
\title{Add a STOPSIGNAL instruction to a dockerfile}
\usage{
dfi_stopsignal(dockerfile, signal)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{signal}{Signal for stopping container}
}
\value{
Updated dockerfile object
}
\description{
Add a STOPSIGNAL instruction to a dockerfile
}
