% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-modifications.R
\name{dfm_add_line}
\alias{dfm_add_line}
\title{Add a line to a dockerfile at a specific position}
\usage{
dfm_add_line(dockerfile, line, after = NULL)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{line}{Line to add}

\item{after}{Position after which to add the line (default: end)}
}
\value{
Updated dockerfile object
}
\description{
Add a line to a dockerfile at a specific position
}
\keyword{internal}
