% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-modifications.R
\name{dfm_group_similar}
\alias{dfm_group_similar}
\title{Group similar instructions in a dockerfile}
\usage{
dfm_group_similar(dockerfile)
}
\arguments{
\item{dockerfile}{A dockerfile object}
}
\value{
Updated dockerfile object with similar instructions grouped
}
\description{
Group similar instructions in a dockerfile
}
