% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-modifications.R
\name{dfm_replace_line}
\alias{dfm_replace_line}
\title{Replace a line in a dockerfile}
\usage{
dfm_replace_line(dockerfile, line_num, new_line)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{line_num}{Line number to replace}

\item{new_line}{New line content}
}
\value{
Updated dockerfile object
}
\description{
Replace a line in a dockerfile
}
