% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-instructions.R
\name{di_add}
\alias{di_add}
\title{Add patterns to a dockerignore object}
\usage{
di_add(dockerignore, pattern)
}
\arguments{
\item{dockerignore}{A dockerignore object}

\item{pattern}{Character vector of patterns to add}
}
\value{
Updated dockerignore object
}
\description{
Add patterns to a dockerignore object
}
