% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-instructions.R
\name{di_remove}
\alias{di_remove}
\title{Remove patterns from a dockerignore object}
\usage{
di_remove(dockerignore, pattern)
}
\arguments{
\item{dockerignore}{A dockerignore object}

\item{pattern}{Character vector of patterns to remove}
}
\value{
Updated dockerignore object
}
\description{
Remove patterns from a dockerignore object
}
