% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-instructions.R
\name{di_replace}
\alias{di_replace}
\title{Replace patterns in a dockerignore object}
\usage{
di_replace(dockerignore, old_pattern, new_pattern)
}
\arguments{
\item{dockerignore}{A dockerignore object}

\item{old_pattern}{Pattern(s) to replace}

\item{new_pattern}{New pattern(s)}
}
\value{
Updated dockerignore object
}
\description{
Replace patterns in a dockerignore object
}
