% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-from-environment.R
\name{dk_from_description}
\alias{dk_from_description}
\title{Create a dockerfile from a DESCRIPTION file}
\usage{
dk_from_description(
  description_file = "DESCRIPTION",
  r_version = NULL,
  base_image = NULL,
  include_sysreqs = TRUE
)
}
\arguments{
\item{description_file}{Path to DESCRIPTION file}

\item{r_version}{R version to use (default: from DESCRIPTION)}

\item{base_image}{Base image to use (default: determined from R version)}

\item{include_sysreqs}{Include system requirements (default: TRUE)}
}
\value{
A dockerfile object
}
\description{
Create a dockerfile from a DESCRIPTION file
}
