% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-from-environment.R
\name{dk_from_script}
\alias{dk_from_script}
\title{Create a dockerfile from an R script}
\usage{
dk_from_script(
  script_file,
  base_image = "rocker/r-ver:latest",
  include_sysreqs = TRUE
)
}
\arguments{
\item{script_file}{Path to R script}

\item{base_image}{Base image to use (default: latest rocker/r-ver)}

\item{include_sysreqs}{Include system requirements (default: TRUE)}
}
\value{
A dockerfile object
}
\description{
Create a dockerfile from an R script
}
