% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-templates.R
\name{dk_register_template}
\alias{dk_register_template}
\title{Register a custom dockerfile template}
\usage{
dk_register_template(name, template_fn)
}
\arguments{
\item{name}{Template name}

\item{template_fn}{Function that returns a dockerfile}
}
\value{
Invisible TRUE if successful
}
\description{
Register a custom dockerfile template
}
