% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-templates.R
\name{dk_template_base}
\alias{dk_template_base}
\title{Create a base R dockerfile template}
\usage{
dk_template_base(r_version = NULL, additional_pkgs = NULL)
}
\arguments{
\item{r_version}{R version to use (default: current version)}

\item{additional_pkgs}{Additional R packages to install}
}
\value{
A dockerfile object
}
\description{
Create a base R dockerfile template
}
