% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-templates.R
\name{dk_template_custom}
\alias{dk_template_custom}
\title{Create a dockerfile from a custom template}
\usage{
dk_template_custom(template_name, ...)
}
\arguments{
\item{template_name}{Name of the template}

\item{...}{Arguments to pass to the template function}
}
\value{
A dockerfile object
}
\description{
Create a dockerfile from a custom template
}
