% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_editor}
\alias{dk_template_ignore_editor}
\title{Create a dockerignore template for editor-related files}
\usage{
dk_template_ignore_editor()
}
\value{
A dockerignore object with editor-related ignore patterns
}
\description{
Create a dockerignore template for editor-related files
}
