% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_os}
\alias{dk_template_ignore_os}
\title{Create a dockerignore template for OS-related files}
\usage{
dk_template_ignore_os(.dockerignore = NULL)
}
\arguments{
\item{.dockerignore}{Optional existing dockerignore object to add patterns to}
}
\value{
A dockerignore object with OS-related ignore patterns
}
\description{
Create a dockerignore template for OS-related files
}
