% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_packrat}
\alias{dk_template_ignore_packrat}
\title{Create a \code{dockerignore} template for \code{{packrat}}-related files}
\usage{
dk_template_ignore_packrat(.dockerignore = NULL)
}
\arguments{
\item{.dockerignore}{Optional existing \code{dockerignore} object to add patterns to}
}
\value{
A \code{dockerignore} object with \code{{packrat}}-related ignore patterns
}
\description{
Creates a \code{dockerignore} template with patterns to ignore \code{{packrat}} library and
source directories.
}
\details{
This template adds patterns to ignore \code{{packrat}} library and source directories,
which are typically not needed in a Docker image. These include:
\itemize{
\item \verb{packrat/lib*/}: Package libraries
\item \verb{packrat/src/}: Package sources
}

When using \code{{packrat}} in a \strong{Dockerfile}, you typically want to copy just the
\code{{packrat}} configuration files and use \code{{packrat}}'s restore functionality to
rebuild the library inside the container, rather than copying the entire
library.
}
\examples{
# Create a new dockerignore with packrat patterns
di <- dk_template_ignore_packrat()

# Add packrat patterns to an existing dockerignore
di <- dockerignore() |>
  di_add("*.log") |>
  dk_template_ignore_packrat()

}
\seealso{
\code{\link[=dk_template_ignore_renv]{dk_template_ignore_renv()}} for renv-specific patterns,
\code{\link[=dk_template_ignore_r]{dk_template_ignore_r()}} for R-specific patterns, &
\code{\link[=dk_template_ignore_common]{dk_template_ignore_common()}} for a more comprehensive template

Other dockerignore template functions: 
\code{\link{dk_template_ignore_common}()},
\code{\link{dk_template_ignore_data}()},
\code{\link{dk_template_ignore_editor}()},
\code{\link{dk_template_ignore_git}()},
\code{\link{dk_template_ignore_node}()},
\code{\link{dk_template_ignore_os}()},
\code{\link{dk_template_ignore_python}()},
\code{\link{dk_template_ignore_r}()},
\code{\link{dk_template_ignore_raw_data}()},
\code{\link{dk_template_ignore_renv}()}
}
\concept{dockerignore template functions}
