% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_python}
\alias{dk_template_ignore_python}
\title{Create a dockerignore template for Python projects}
\usage{
dk_template_ignore_python()
}
\value{
A dockerignore object with Python-related ignore patterns
}
\description{
Create a dockerignore template for Python projects
}
