% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_raw_data}
\alias{dk_template_ignore_raw_data}
\title{Create a dockerignore template for raw data directories}
\usage{
dk_template_ignore_raw_data()
}
\value{
A dockerignore object with raw data directory patterns
}
\description{
Create a dockerignore template for raw data directories
}
