% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_renv}
\alias{dk_template_ignore_renv}
\title{Create a dockerignore template for renv-related files}
\usage{
dk_template_ignore_renv(.dockerignore = NULL)
}
\arguments{
\item{.dockerignore}{Optional existing dockerignore object to add patterns to}
}
\value{
A dockerignore object with renv-related ignore patterns
}
\description{
Create a dockerignore template for renv-related files
}
