% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-templates.R
\name{dk_template_plumber}
\alias{dk_template_plumber}
\title{Create a Plumber API dockerfile template}
\usage{
dk_template_plumber(
  r_version = NULL,
  port = 8000,
  api_file = "plumber.R",
  additional_pkgs = NULL
)
}
\arguments{
\item{r_version}{R version to use (default: current version)}

\item{port}{Port to expose (default: 8000)}

\item{api_file}{Path to plumber.R file (default: "plumber.R")}

\item{additional_pkgs}{Additional R packages to install}
}
\value{
A dockerfile object
}
\description{
Create a Plumber API dockerfile template
}
