% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pkg_install_cmd}
\alias{generate_pkg_install_cmd}
\title{Generate package installation commands for different package managers}
\usage{
generate_pkg_install_cmd(package_manager, packages)
}
\arguments{
\item{package_manager}{Character string specifying the package manager to use}

\item{packages}{Character vector of packages to install}
}
\value{
Character vector of installation commands
}
\description{
Generate package installation commands for different package managers
}
