% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_manager}
\alias{get_package_manager}
\title{Detect the package manager from a base image}
\usage{
get_package_manager(base_image)
}
\arguments{
\item{base_image}{Base image name}
}
\value{
Character string of package manager type ("apt", "yum", "apk", etc.)
}
\description{
Detect the package manager from a base image
}
