% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-core.R
\name{is_dockerignore}
\alias{is_dockerignore}
\title{Test if an object is a dockerignore}
\usage{
is_dockerignore(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
\code{TRUE} if \code{x} is a dockerignore object, \code{FALSE} otherwise
}
\description{
Checks whether the provided object is a valid \code{dockerignore} class object.
}
\examples{
di <- dockerignore()
is_dockerignore(di)  # TRUE
is_dockerignore(list())  # FALSE

}
\seealso{
\code{\link[=dockerignore]{dockerignore()}} for creating a dockerignore object &
\code{\link[=check_dockerignore]{check_dockerignore()}} for ensuring an object is a dockerignore (with error)

Other dockerignore core functions: 
\code{\link{c.dockerignore}()},
\code{\link{check_dockerignore}()},
\code{\link{dockerignore}()},
\code{\link{print.dockerignore}()}
}
\concept{dockerignore core functions}
