% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_to_sysreqs_platform}
\alias{map_to_sysreqs_platform}
\title{Map package manager to sysreqs platform}
\usage{
map_to_sysreqs_platform(package_manager, os = NULL)
}
\arguments{
\item{package_manager}{Package manager type}

\item{os}{Operating system (if known)}
}
\value{
Character string of platform for sysreqs
}
\description{
Map package manager to sysreqs platform
}
\keyword{internal}
