% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-io.R
\name{read_dockerfile}
\alias{read_dockerfile}
\title{Read a Dockerfile from a file}
\usage{
read_dockerfile(file)
}
\arguments{
\item{file}{Path to Dockerfile}
}
\value{
A dockerfile object
}
\description{
Read a Dockerfile from a file
}
