% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-io.R
\name{read_dockerignore}
\alias{read_dockerignore}
\title{Read a .dockerignore file}
\usage{
read_dockerignore(file = ".dockerignore")
}
\arguments{
\item{file}{Path to .dockerignore file}
}
\value{
A dockerignore object
}
\description{
Read a .dockerignore file
}
