% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-templates.R
\docType{data}
\name{template_registry}
\alias{template_registry}
\title{Templates registry for storing custom templates}
\format{
An object of class \code{environment} of length 0.
}
\usage{
template_registry
}
\description{
Templates registry for storing custom templates
}
\keyword{internal}
