% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-io.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Write a dockerfile to a file}
\usage{
write_dockerfile(dockerfile, file = "Dockerfile", multiline = TRUE)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{file}{Output file path (default: "Dockerfile")}

\item{multiline}{Logical indicating if long RUN commands should be split (default: TRUE)}
}
\value{
Invisible dockerfile object
}
\description{
Write a dockerfile to a file
}
