% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-io.R
\name{write_dockerignore}
\alias{write_dockerignore}
\title{Write a dockerignore object to a file}
\usage{
write_dockerignore(dockerignore, file = ".dockerignore")
}
\arguments{
\item{dockerignore}{A dockerignore object}

\item{file}{Output file path (default: ".dockerignore")}
}
\value{
Invisible dockerignore object
}
\description{
Write a dockerignore object to a file
}
