% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-core.R
\name{add_dockerfile_line}
\alias{add_dockerfile_line}
\title{Add a line to a dockerfile and update metadata}
\usage{
add_dockerfile_line(dockerfile, instruction, args)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{instruction}{Docker instruction (e.g., "FROM", "RUN")}

\item{args}{Arguments for the instruction}
}
\value{
Updated dockerfile object
}
\description{
Add a line to a dockerfile and update metadata
}
\keyword{internal}
