% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-core.R
\name{c.dockerignore}
\alias{c.dockerignore}
\title{Combine multiple dockerignore objects}
\usage{
\method{c}{dockerignore}(...)
}
\arguments{
\item{...}{Dockerignore objects to combine}
}
\value{
A new dockerignore object with combined patterns
}
\description{
Combine multiple dockerignore objects
}
