% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_add}
\alias{dfi_add}
\title{Add an ADD instruction to a dockerfile}
\usage{
dfi_add(dockerfile, src, dest)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{src}{Source path (can be a URL)}

\item{dest}{Destination path}
}
\value{
Updated dockerfile object
}
\description{
Add an ADD instruction to a dockerfile
}
