% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_arg}
\alias{dfi_arg}
\title{Add an ARG instruction to a dockerfile}
\usage{
dfi_arg(dockerfile, name, default = NULL)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{name}{Argument name}

\item{default}{Default value (optional)}
}
\value{
Updated dockerfile object
}
\description{
Add an ARG instruction to a dockerfile
}
