% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_copy}
\alias{dfi_copy}
\title{Add a COPY instruction to a dockerfile}
\usage{
dfi_copy(dockerfile, src, dest, from = NULL)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{src}{Source path}

\item{dest}{Destination path}

\item{from}{Build stage to copy from (optional)}
}
\value{
Updated dockerfile object
}
\description{
Add a COPY instruction to a dockerfile
}
