% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_env}
\alias{dfi_env}
\title{Add an ENV instruction to a dockerfile}
\usage{
dfi_env(dockerfile, ...)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{...}{Named environment variables}
}
\value{
Updated dockerfile object
}
\description{
Add an ENV instruction to a dockerfile
}
