% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_run}
\alias{dfi_run}
\title{Add a RUN instruction to a dockerfile}
\usage{
dfi_run(dockerfile, commands)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{commands}{Commands to run}
}
\value{
Updated dockerfile object
}
\description{
Add a RUN instruction to a dockerfile
}
