% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_user}
\alias{dfi_user}
\title{Add a USER instruction to a dockerfile}
\usage{
dfi_user(dockerfile, user, group = NULL)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{user}{Username or UID}

\item{group}{Group or GID (optional)}
}
\value{
Updated dockerfile object
}
\description{
Add a USER instruction to a dockerfile
}
