% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_volume}
\alias{dfi_volume}
\title{Add a VOLUME instruction to a dockerfile}
\usage{
dfi_volume(dockerfile, paths)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{paths}{Path(s) to create as volumes}
}
\value{
Updated dockerfile object
}
\description{
Add a VOLUME instruction to a dockerfile
}
