% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-instructions.R
\name{dfi_workdir}
\alias{dfi_workdir}
\title{Add a WORKDIR instruction to a dockerfile}
\usage{
dfi_workdir(dockerfile, path)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{path}{Working directory path}
}
\value{
Updated dockerfile object
}
\description{
Add a WORKDIR instruction to a dockerfile
}
