% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-modifications.R
\name{dfm_move_line}
\alias{dfm_move_line}
\title{Move a line in a dockerfile}
\usage{
dfm_move_line(dockerfile, from, to)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{from}{Source line number}

\item{to}{Target position}
}
\value{
Updated dockerfile object
}
\description{
Move a line in a dockerfile
}
