% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-modifications.R
\name{dfm_remove_line}
\alias{dfm_remove_line}
\title{Remove a line from a dockerfile}
\usage{
dfm_remove_line(dockerfile, line_num)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{line_num}{Line number to remove}
}
\value{
Updated dockerfile object
}
\description{
Remove a line from a dockerfile
}
