% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-modifications.R
\name{dfm_sort_by_instruction}
\alias{dfm_sort_by_instruction}
\title{Sort instructions in a dockerfile by type}
\usage{
dfm_sort_by_instruction(dockerfile, order = NULL)
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{order}{Custom order for instructions (optional)}
}
\value{
Updated dockerfile with sorted instructions
}
\description{
Sort instructions in a dockerfile by type
}
