% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-from-environment.R
\name{dk_add_sysreqs}
\alias{dk_add_sysreqs}
\title{Add system requirements for R packages to a dockerfile}
\usage{
dk_add_sysreqs(dockerfile, packages, package_manager = "auto")
}
\arguments{
\item{dockerfile}{A dockerfile object}

\item{packages}{Character vector of package names}

\item{package_manager}{Package manager to use (default: auto-detected)}
}
\value{
Updated dockerfile object
}
\description{
Add system requirements for R packages to a dockerfile
}
