% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockitect-from-environment.R
\name{dk_from_renv}
\alias{dk_from_renv}
\title{Create a dockerfile from an renv.lock file}
\usage{
dk_from_renv(
  lock_file = "renv.lock",
  r_version = NULL,
  base_image = NULL,
  include_sysreqs = TRUE
)
}
\arguments{
\item{lock_file}{Path to renv.lock file}

\item{r_version}{R version to use (default: from lock file)}

\item{base_image}{Base image to use (default: determined from R version)}

\item{include_sysreqs}{Include system requirements (default: TRUE)}
}
\value{
A dockerfile object
}
\description{
Create a dockerfile from an renv.lock file
}
