% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_data}
\alias{dk_template_ignore_data}
\title{Create a dockerignore template for common data files}
\usage{
dk_template_ignore_data(include_raw = TRUE)
}
\arguments{
\item{include_raw}{Include raw data directories (default: TRUE)}
}
\value{
A dockerignore object with data-related ignore patterns
}
\description{
Create a dockerignore template for common data files
}
