% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_git}
\alias{dk_template_ignore_git}
\title{Create a dockerignore template for Git-related files}
\usage{
dk_template_ignore_git(.dockerignore = NULL)
}
\arguments{
\item{.dockerignore}{Optional existing dockerignore object to add patterns to}
}
\value{
A dockerignore object with Git-related ignore patterns
}
\description{
Create a dockerignore template for Git-related files
}
