% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_node}
\alias{dk_template_ignore_node}
\title{Create a dockerignore template for Node.js/JavaScript projects}
\usage{
dk_template_ignore_node()
}
\value{
A dockerignore object with Node.js-related ignore patterns
}
\description{
Create a dockerignore template for Node.js/JavaScript projects
}
