% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_os}
\alias{dk_template_ignore_os}
\title{Create a dockerignore template for OS-related files}
\usage{
dk_template_ignore_os()
}
\value{
A dockerignore object with OS-related ignore patterns
}
\description{
Create a dockerignore template for OS-related files
}
