% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_packrat}
\alias{dk_template_ignore_packrat}
\title{Create a dockerignore template for packrat-related files}
\usage{
dk_template_ignore_packrat()
}
\value{
A dockerignore object with packrat-related ignore patterns
}
\description{
Create a dockerignore template for packrat-related files
}
