% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-templates.R
\name{dk_template_ignore_r}
\alias{dk_template_ignore_r}
\title{Create a dockerignore template for R-related files}
\usage{
dk_template_ignore_r(
  .dockerignore = NULL,
  include_renv = TRUE,
  include_packrat = TRUE
)
}
\arguments{
\item{.dockerignore}{Optional existing dockerignore object to add patterns to}

\item{include_renv}{Include renv folders (default: TRUE)}

\item{include_packrat}{Include packrat folders (default: TRUE)}
}
\value{
A dockerignore object with R-related ignore patterns
}
\description{
Create a dockerignore template for R-related files
}
