% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-core.R
\name{is_dockerfile}
\alias{is_dockerfile}
\title{Test if an object is a dockerfile}
\usage{
is_dockerfile(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
TRUE if x is a dockerfile, FALSE otherwise
}
\description{
Test if an object is a dockerfile
}
