% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-core.R
\name{is_dockerignore}
\alias{is_dockerignore}
\title{Test if an object is a dockerignore}
\usage{
is_dockerignore(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
TRUE if x is a dockerignore, FALSE otherwise
}
\description{
Test if an object is a dockerignore
}
