% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerignore-core.R
\name{print.dockerignore}
\alias{print.dockerignore}
\title{Print method for dockerignore objects}
\usage{
\method{print}{dockerignore}(x, ...)
}
\arguments{
\item{x}{A dockerignore object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for dockerignore objects
}
